pyproject.toml
[project]
name = "autocourse-ai"
version = "0.1.0"
description = "AI-powered multi-modal course generation system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # DSPy
    "dspy-ai>=2.5.0",
    # Data validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    # Vector store
    "chromadb>=0.5.0",
    # Embeddings
    "sentence-transformers>=3.0.0",
    # PDF extraction
    "pymupdf>=1.24.0",
    # YouTube ingestion
    "yt-dlp>=2024.11.0",
    # Transcription
    "openai-whisper>=20240930",
    # Utilities
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "anyio>=4.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "anyio>=4.6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-report=html:reports/coverage -v"

[tool.hatch.build.targets.wheel]
packages = ["app"]
